prepare reading messages

+++ src/backend/Bombi/Bombi.ServerInstance/Game/Constants.cs
@@ -5,7 +5,7 @@ public static class Constants
     public const int LevelWidth = 15;
     public const int LevelHeight = 13;
 
-    public const double BoxSpawnPropability = 0.9;
+    public const float BoxSpawnPropability = 0.9f;
 
-    public const double TileSize = 32.0;
+    public const float TileSize = 32.0f;
 }
